# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API BASE",
        "description": "Servicio API REST para la gestión y administración del sistema Automatizado de riego por goteo",
        "contact": {
            "name": "SETI",
            "email": "cirojas@cenicana.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/irrigation-headers": {
            "get": {
                "tags": [
                    "IrrigationHeaders"
                ],
                "summary": "Get list of irrigation headers",
                "description": "Returns list of irrigation headers",
                "operationId": "getIrrigationHeaderList",
                "parameters": [
                    {
                        "name": "estate_id",
                        "in": "query",
                        "description": "The ID of the estate (hacienda)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityJsonResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "IrrigationHeaders"
                ],
                "summary": "Store new irrigation header",
                "description": "Returns irrigation header data",
                "operationId": "storeIrrigationHeader",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IrrigationHeader"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IrrigationHeader"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
        "/v1/irrigation-headers/{id}": {
            "get": {
                "tags": [
                    "IrrigationHeaders"
                ],
                "summary": "Get irrigation header information",
                "description": "Returns irrigation header data",
                "operationId": "getIrrigationHeadersById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Irrigation header id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityJsonResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Login",
                "description": "Genera el token de autenticación",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "E-Mail",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Contraseña",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "El login se realizó correctamente "
                    },
                    "400": {
                        "description": "Petición no valida"
                    },
                    "401": {
                        "description": "Usuario no autorizado"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Logout",
                "description": "Invalida el token",
                "operationId": "logout",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "El login se realizó correctamente "
                    },
                    "400": {
                        "description": "Petición no valida"
                    },
                    "401": {
                        "description": "Usuario no autorizado"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Refresca el token de autenticación",
                "description": "Invalida el token actual y genera uno nuevo",
                "operationId": "refresh",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "El login se realizó correctamente "
                    },
                    "400": {
                        "description": "Petición no valida"
                    },
                    "401": {
                        "description": "Usuario no autorizado"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtiene la información del usuaro autenticado",
                "description": "Obtiene la información del usuaro autenticado",
                "operationId": "me",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "El login se realizó correctamente "
                    },
                    "400": {
                        "description": "Petición no valida"
                    },
                    "401": {
                        "description": "Usuario no autorizado"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "IrrigationHeader": {
                "title": "Irrigation Header model",
                "description": "IrrigationHeader model",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the irrigation header",
                        "type": "string",
                        "example": "Cabezal A"
                    },
                    "estate_id": {
                        "title": "EstateId",
                        "description": "The ID of the estate (Hacienda)",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "water_source_id": {
                        "title": "WaterSourceId",
                        "description": "The ID of the water source (Fuente de agua)",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "motorpump_brand": {
                        "title": "Motorpump brand",
                        "description": "The brand of the motorpump (la marca de la motobomba)",
                        "type": "string",
                        "example": "Honda"
                    },
                    "motorpump_reference": {
                        "title": "Motorpump reference",
                        "description": "The reference of the motorpump (la ref de la motobomba)",
                        "type": "string",
                        "example": "modelo XYZ"
                    },
                    "motorpump_hp": {
                        "title": "Motorpump HP",
                        "description": "The HP of the motorpump (caballos de fuerza de la motobomba)",
                        "type": "number",
                        "format": "float",
                        "example": "10.5"
                    },
                    "motorpump_flow": {
                        "title": "Motorpump flow",
                        "description": "The flow of the motorpump (flujo de la motobomba)",
                        "type": "number",
                        "format": "float",
                        "example": "5.5"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "IrrigationHeader"
                }
            },
...
            "EntityJsonResource": {
                "title": "EntityJsonResource",
                "description": "A Json wrapper resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "ref": null
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EntityJsonResource"
                }
            }
        },
        "securitySchemes": {
            "jwt_key": {
                "type": "apiKey",
                "description": "Debe insertar el token generado en el proceso de autenticación",
                "name": "Authorization",
                "in": "header",
                "template": "Bearer {}"
            }
        }
    },
...
}